Build fix: Get rid of second call to max
By changing the type of the member variable _timeout_count to uint32_t
the likelyhood of this wrapping around is minimized so we can remove
the second call to max that failed the build.

So lets remove the call to max in order to be able to compile this
great library. As a side note the call to max looks like it should
have been a call to min.